Render test results using html2!  break embox...
This commit does the minimum to switch to the new renderer,
by mapping .html to html2_scanner.  Now we have to remove
all the references to "2" and all the old dead code, of course.

Our use of emboxValue breaks test-in-browser pretty severely
by assuming that in `{{#each results}}`, if `results` is `===`
to what it was last time, then there is no change.  In fact,
test-in-browser mutates arrays and expects invalidating the
result to cause a re-render.