Implement new # and > calling convention
Both {{#foo}} and {{> foo}} are handled as follows:

- foo may be a template or a function returning a template
- if there are only keyword arguments, they are used to create a data context object (e.g. `x=1 y=2` becomes `{x:1, y:2}`)
- if there are any positional arguments, the arguments are treated as a nested helper call, meaning that the first one will be called on the rest (including keyword arguments) if it is a function

Rendering-wise, the template should only be re-rendered if foo is a function and it is invalidated to return a different value (though the "isolate" that would compare the new value to the old is not on this branch at the moment).

This change also makes the built-in block helpers (#if, #each, etc.) be functions (render macros) instead of components (though #each is still implemented with a component).  This avoids data context issues with #if and #unless (the new calling convention isn't really designed to support making something like #if).  It also should make generated code cleaner to call `UI.If` instead of including it as a component.