Only look for helpers in the current template
Revert the new feature that all helpers in enclosing templates are visible, including helpers on the template that called the current template (for example) and helpers on UI.body.

Making helpers visible to sub-templates was a major change that was never properly justified or thought through.  Helpers take precedence over data properties, and having helpers "leak" down through the app could have bad consequences, even if it's useful for defining helpers on a group of related templates.  There was a point in shark's history where we needed this dynamic lookup in order for helpers to work in block tags, but that's no longer the case.

Now, helpers are only visible to the lexical body of the template they are defined on (except for global helpers defined with UI.registerHelper).

Ran tests in Chrome; ran todos and docs.