Correctly unbind focus and blur events
With jQuery, you can’t do `$elem.on(‘focus’, ‘.selector’, handler)` and then `$elem.off(‘focus’, handler)`.  The correct code is `$elem.off(‘focus’, ‘**’, handler)`, in order to remove the handler without having to specify the selector.

The reason is that `on` with a selector actually listens to focusin/focusout, not focus/blur, and then `off` tries to remove focus/blur.