Fix Parties on IE9 by not using parentElement
Using parentElement instead of parentNode was probably a typo.  I didn’t even know there was an Element#parentElement alongside Node#parentNode in the DOM, but apparently there is, and apparently IE 9 supports it, but for some reason it doesn’t work.

The lesson is: never use parentElement on a DOM node.  Use parentNode and be done with it.