Don't try to uglify on the client
Follow-up to 36ff10c08.

Note that 696ce39c101 made tests fail, because it used minifiers on the
client for the first time, which hit this edge case.

If you are hitting this issue in your own app because you are using
SpacebarsCompiler on the client and loading minifiers, you can work
around it for now by putting this somewhere:

   Package.minifiers.UglifyJSMinify = function (code) {
     return { code: code };
   };