Added API box from meteor docs
It's kind of ugly to copy this much stuff over, I'd like to make an
jsdoc_api_box hexo plugin, but there's so much specific stuff in here
that needs to be factored out properly (mainly the jsdoc/ scripts and
config).

Maybe some can do it some day but for now we'll have some duplication.
Hopefully we won't need to change it much.