Resources Link Update
Templates API and Blaze API links directed to the Meteor Guide which then redirects back to a page on blazejs.org. Those resource links have been updated to direct straight to the blazejs.org pages.

The Spacebars link directs to a 404 on Github and there is Spacebars package documentation on Meteor docs, so that link has been updated to the docs.