Update Hexo version and introduce package-lock.json for docs.
While Hexo 3.3.9 isn't the latest version, it is the latest version which
doesn't incorrectly assume that `hexo-s3-deploy` is a Hexo plugin (it's
not).

This also re-introduces a value for `package.json:hexo.version`, which was
removed in 673d239660ef28c3ff1436e0e7c2cfb38c14bea6.  Unfortunately, it must
stay as a defined value, as per Hexo's demands.